# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "description": "Returns the caller's own categories together with the shared catalogue, sorted by name.\n\nEach entry reports whether it is `shared`. A shared category can be selected by any account but cannot be renamed, archived, or deleted, so a client should not offer those actions for it.\n\nArchived categories are excluded by default, because a picker wants only what can be chosen now. Pass `includeArchived=true` to see them; they still resolve in existing plans, expenses, and reports.",
        "operationId": "CategoriesController_list_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "How many records to return.",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "How many records to skip.",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include categories that have been archived. Off by default, since a picker wants only selectable ones.",
            "schema": {
              "default": false,
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of selectable categories.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List selectable categories",
        "tags": [
          "Categories"
        ]
      },
      "post": {
        "description": "Creates a category owned by the caller, alongside the shared catalogue.\n\nNames are compared case and spacing insensitively, so \"Cloud Hosting\" and \"cloud  hosting\" collide and the second is rejected. Without that, a picker ends up showing what looks like the same category twice and a report splits the same spend across both.\n\nCreating a category with the same name as a shared one is allowed. The account's own category takes precedence when a CSV import resolves that name.",
        "operationId": "CategoriesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "409": {
            "description": "The caller already has a category with that name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{categoryId}": {
      "patch": {
        "description": "Changes the name, the archived state, or both. Anything omitted is left alone.\n\nOnly the caller's own categories can be changed. A shared catalogue entry answers 404 rather than explaining the rule, because a client should already know from the `shared` flag and there is nothing useful to reveal.\n\nArchiving hides a category from pickers without affecting anything historic: existing plans and expenses keep resolving it and reports are unchanged. It is almost always what someone wants instead of deleting.",
        "operationId": "CategoriesController_update_v1",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "description": "Identifier from the category list.",
            "schema": {
              "example": "65f1c2d3e4b5a6c7d8e9f0a1",
              "allOf": [
                {
                  "$ref": "#/components/schemas/ObjectId"
                }
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated category.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "The caller does not own a category with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "409": {
            "description": "The new name collides with another of the caller’s categories.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename or archive a category",
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "description": "Removes a category the caller owns from every list.\n\nThe delete is soft. The record survives so that plans and expenses in a locked period keep resolving the name they referenced, and so a mistaken delete is recoverable without a database restore. The change is recorded in the audit log with the category's last state.\n\nConsider archiving instead. Deleting a category that already has spend against it makes those records harder to read, while archiving hides it from pickers and leaves history intact.\n\nA shared catalogue entry answers 404.",
        "operationId": "CategoriesController_remove_v1",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "description": "Identifier from the category list.",
            "schema": {
              "example": "65f1c2d3e4b5a6c7d8e9f0a1",
              "allOf": [
                {
                  "$ref": "#/components/schemas/ObjectId"
                }
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The category was deleted."
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "The caller does not own a category with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a category",
        "tags": [
          "Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plan vs Actual API",
    "description": "Monthly spending targets, logged expenses, and variance reporting with locked periods.\n\n**Money** is an integer count of minor units. A field ending in `Minor` holds cents.\n\n**Months** are the string `YYYY-MM`.\n\n**Variance** is `actual - plan`. Variance percent is `null` when the plan is zero, never `NaN`.\n\n**Missing actuals** default to `0`. Pass `missingActuals=null` to receive `null` instead. Every report row carries `hasActual`, so a logged zero is never confused with nothing logged.\n\n**Locked periods** reject writes with `423` and the code `PERIOD_LOCKED`.\n\n**Deletes are soft.** A `DELETE` answers `204` and the record stops appearing in reads.\n\n**Sessions** travel in the `Authorization` header, never a cookie, so a mobile or desktop client is a first class caller. Access tokens are RS256 signed and short lived; refresh tokens are opaque, single use, and rotate.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponseDTO": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status, repeated in the body.",
            "example": 423
          },
          "code": {
            "type": "string",
            "description": "Stable code a client branches on. Never parse the message.",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "IMPORT_VALIDATION_FAILED",
              "PERIOD_LOCKED",
              "RATE_LIMITED",
              "INTERNAL"
            ],
            "example": "PERIOD_LOCKED"
          },
          "message": {
            "type": "string",
            "description": "What went wrong, written for a person.",
            "example": "2026-01 is locked and cannot be edited."
          },
          "details": {
            "type": "object",
            "description": "Machine readable specifics: the offending month, the failing rows, the invalid fields.",
            "example": {
              "month": "2026-01"
            }
          },
          "requestId": {
            "type": "string",
            "description": "Correlates this response with its log lines and the x-request-id header.",
            "example": "0f9c1e2a-7b3d-4c5e-9a8f-1b2c3d4e5f6a"
          },
          "path": {
            "type": "string",
            "description": "The path that produced the error.",
            "example": "/api/v1/plans"
          },
          "timestamp": {
            "type": "string",
            "description": "When the error was produced.",
            "example": "2026-01-15T10:04:11.212Z",
            "format": "date-time"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "requestId",
          "path",
          "timestamp"
        ]
      },
...
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
...
      "CategoryResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier, used when creating a plan or an expense.",
            "example": "65f1c2d3e4b5a6c7d8e9f0a1"
          },
          "name": {
            "type": "string",
            "description": "The name as written.",
            "example": "Cloud Hosting"
          },
          "shared": {
            "type": "boolean",
            "description": "Whether this comes from the shared catalogue. Shared categories are readable by every account and editable by none.",
            "example": false
          },
          "archived": {
            "type": "boolean",
            "description": "Whether it is hidden from pickers. An archived category still resolves in existing plans and reports.",
            "example": false
          }
        },
        "required": [
          "id",
          "name",
          "shared",
          "archived"
        ]
      },
      "CreateCategoryDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name as it should be displayed. Uniqueness is checked case and spacing insensitively, so \"Cloud Hosting\" and \"cloud  hosting\" are the same category.",
            "example": "Cloud Hosting",
            "minLength": 1,
            "maxLength": 60
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "A new display name.",
            "example": "Cloud Infrastructure",
            "minLength": 1,
            "maxLength": 60
          },
          "archived": {
            "type": "boolean",
            "description": "Hide the category from pickers without removing it. Existing plans and expenses keep resolving it, and reports are unaffected.",
            "example": true
          }
        }
      },
...
    }
  }
}